backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-groq
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
slack_sdk
notion-client
PyGithub
requests
pypdf
